backend/requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
flask-cors>=5.0
psycopg2-binary>=2.9
python-dotenv>=1.0
PyJWT>=2.8
cryptography>=42.0
requests>=2.31
